+++ src/Templates/Template.php
@@ -43,7 +43,7 @@ abstract class Template implements TemplateInterface
     /**
      * string VendorVersion.
      */
-    protected $ver = '<b>Ver:</b> dev.2605.1800';
+    protected string $ver = '<b>Ver:</b> dev.2608.1700';
 
     /**
      * TemplateDefault constructor.